move components folder related to posts under posts folder in views
