+++ frontend/src/app/app.routes.ts
@@ -3,7 +3,7 @@ import { Routes, RouterModule } from '@angular/router';
 import { LoginComponent } from './views/login/login.component';
 import { RegisterComponent } from './views/register/register.component';
 import { MainComponent } from './views/main/main.component';
-import { DisplayPostComponent } from './components/display-post/display-post.component';
+import { DisplayPostComponent } from './views/post/components/display-post/display-post.component';
 
 export const routes: Routes = [
   { path: '', component: MainComponent },
